Refactor step imports in ZenML components
